Refactor and change enable retry by default
