+++ .generator/src/generator/templates/utils.j2
@@ -2,7 +2,10 @@
 package {{ common_package_name }}
 
 import (
+	"bytes"
 	"encoding/json"
+	"io/ioutil"
+	"net/http"
 	"reflect"
 	"strings"
 	"time"
@@ -442,4 +445,15 @@ func ContainsUnparsedObject(i interface{}) (bool, interface{}) {
 // Strlen returns number of runes in string
 func Strlen(s string) int {
 	return utf8.RuneCountInString(s)
+}
+
+// Copy the original request so it doesn't get lost when retrying
+func copyRequest(r *http.Request, rawBody *[]byte) *http.Request {
+	newRequest := *r
+
+	if r.Body == nil || r.Body == http.NoBody {
+		return &newRequest
+	}
+	newRequest.Body = ioutil.NopCloser(bytes.NewBuffer(*rawBody))
+	return &newRequest
 }
